[decompose] extract BlockList and EditProfile from ProfileModalController
