+++ decisions/DECISIONS_2026-02-19_19-58-00.md
@@ -1,0 +1,15 @@
+# Decomposition Decisions
+
+- **File:** `js/ui/profileModalController.js`
+- **Extracted Controllers:**
+    - `js/ui/profileModal/ProfileBlockListController.js` (Block list logic)
+    - `js/ui/profileModal/ProfileEditController.js` (Edit profile logic)
+- **Rationale:**
+    - Block list management is a distinct feature set (add, remove, list, refresh).
+    - Edit profile is a self-contained form handling workflow (populate, upload, save).
+    - Reducing the main controller size improves readability and separation of concerns.
+- **Backwards Compatibility:**
+    - New controllers alias DOM references back to the main controller (`this.mainController.blockList = ...`) to maintain compatibility with tests and potentially other external code that expects these properties on the main controller.
+    - Delegate methods added to `ProfileModalController` to forward calls to the new controllers, preserving the public API.
+- **Baseline Update:**
+    - `js/ui/profileModalController.js` reduced from 6258 to 5541 lines (-717 lines).
