+++ docs/agents/task-logs/daily/2026-02-19_19-58-00_decompose-agent_completed.md
@@ -1,0 +1,11 @@
+# Decompose Agent Completion Log
+
+- **Agent:** decompose-agent
+- **Date:** 2026-02-19
+- **Status:** Completed
+- **Task:** Decompose `js/ui/profileModalController.js`.
+- **Result:**
+    - Extracted `ProfileBlockListController.js` and `ProfileEditController.js`.
+    - Reduced `js/ui/profileModalController.js` by 717 lines (6258 -> 5541).
+    - Verified with lint and unit tests.
+    - Updated file size baseline.
