+++ test_logs/TEST_LOG_2026-02-19_19-58-00.md
@@ -1,0 +1,17 @@
+# Test Log
+
+## Commands Run
+
+1.  `npm run lint`
+    - Result: Passed (ignoring assets check).
+2.  `node scripts/run-targeted-tests.mjs tests/profile-modal-controller.test.mjs`
+    - Result: Passed (18/18 tests).
+    - Initial failure due to missing alias for `this.blockList` on main controller. Fixed by adding aliases in `ProfileBlockListController.js`.
+3.  `node scripts/check-file-size.mjs --report`
+    - Result: `js/ui/profileModalController.js` is now 5541 lines.
+
+## Manual Verification Notes
+
+- Code inspection confirmed extracted logic matches original behavior.
+- Delegate methods ensure API compatibility.
+- Constructor instantiation order preserved (new controllers added before `hashtagController` initialization).
